WIP: First draft of a corpus worker
This is the first draft of a worker that can get a corpus and create an
analysis for it. This first attempt was a freqdist worker, that takes the
freqdist for each document in the corpus and condensates it in a new analysis:
the freqtdist for the entire corpus.

This is a work in progress because I was mainly worried with the basis for this
to work (specially the celery task). I did not pay any attention to the way the
worker itself is working (it's probably doing more work than it needs to), and
it also probably needs more tests.