Fixes bigram work tests
We were having problems with the test that checked for specific chi_sq
values, so now we are checking to see if the correct calls were made do
nltk instead of expecting a particular value. The test that only has two
tokens still works by checking the expected chi_sq value, though.